requirements.txt
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.2
Deprecated==1.2.13
Flask==2.0.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.3
markdown2==2.4.2
MarkupSafe==2.0.1
outcome==1.1.0
pycodestyle==2.8.0
pycparser==2.21
PyGithub==1.55
Pygments==2.12.0
PyJWT==2.3.0
PyNaCl==1.5.0
pynput==1.7.6
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
twilio==7.8.2
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.14.0
wsproto==1.1.0
zipp==3.8.0

# Mac users
# pyobjc-core==8.1
# pyobjc-framework-ApplicationServices==8.1
# pyobjc-framework-Cocoa==8.1
# pyobjc-framework-Quartz==8.1